+++ envisim_estimate/src/nearest_neighbour.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2024 Wilmer Prentius, Anton Grafström.
+// Copyright (C) 2025 Wilmer Prentius, Anton Grafström.
 //
 // This program is free software: you can redistribute it and/or modify it under the terms of the
 // GNU Affero General Public License as published by the Free Software Foundation, version 3.
@@ -12,10 +12,10 @@
 
 //! Nearest neighbour estimator
 
-use envisim_samplr::SamplingError;
-use envisim_utils::kd_tree::{Searcher, TreeBuilder};
+use envisim_samplr::{AuxiliariesOptions, SamplingError};
+use envisim_utils::kd_tree::Searcher;
 use envisim_utils::utils::usize_to_f64;
-use envisim_utils::InputError;
+use envisim_utils::{InputError, Matrix};
 use rustc_hash::{FxBuildHasher, FxHashMap};
 
 /// Nearest neighbour estimator of total.
@@ -23,18 +23,25 @@ use rustc_hash::{FxBuildHasher, FxHashMap};
 pub fn nearest_neighbour(
     y_values: &[f64],
     sample: &[usize],
-    tree_builder: &TreeBuilder,
+    auxiliaries: &Matrix,
 ) -> Result<f64, SamplingError> {
-    let tree = tree_builder.build(&mut sample.to_vec())?;
-    let mut searcher = Searcher::new_1(&tree);
-    let auxilliaries = tree.data();
-
-    let population_size = auxilliaries.nrow();
+    let population_size = auxiliaries.nrow();
     let sample_size = sample.len();
-    InputError::check_lengths(y_values, sample).and(InputError::check_sizes(
+
+    InputError::check_lengths(y_values, sample).and(InputError::check_range_usize(
+        *sample.iter().max().unwrap_or(&0usize),
+        0,
         population_size,
-        auxilliaries.nrow(),
     ))?;
+
+    if sample_size == 0 {
+        return Ok(f64::NAN);
+    }
+
+    let tree = AuxiliariesOptions::new(auxiliaries)?
+        .est_bucket_size()?
+        .build_tree(&mut sample.to_vec())?;
+    let mut searcher = Searcher::new_1(&tree);
 
     let mut number_of_neighbours =
         FxHashMap::<usize, f64>::with_capacity_and_hasher(sample_size, FxBuildHasher);
@@ -45,7 +52,7 @@ pub fn nearest_neighbour(
 
     for i in 0..population_size {
         searcher
-            .find_neighbours_of_iter(&tree, &mut auxilliaries.row_iter(i))
+            .find_neighbours_of_iter(&tree, auxiliaries.row_iter(i))
             .unwrap();
         let part = 1.0 / usize_to_f64(searcher.neighbours().len());